final tweaks

+++ app/src/main/java/pacifier/com/app/events/SpeedChangeEvent.java
@@ -1,10 +1,10 @@
 package pacifier.com.app.events;
 
 public class SpeedChangeEvent {
-    public final float speed;
+    public final int speed;
     public final int speedLimit;
 
-    public SpeedChangeEvent(float speed, int speedLimit) {
+    public SpeedChangeEvent(int speed, int speedLimit) {
         this.speed = speed;
         this.speedLimit = speedLimit;
     }
